+++ Kokoro/APIv2/InPlaceCardUpgrade.cs
@@ -19,12 +19,18 @@ public partial interface IKokoroApi
 			/// <returns>A wrapper, granting access to the modifications.</returns>
 			ICardUpgrade ModifyCardUpgrade(CardUpgrade route);
 			
+			// TODO: XML docs
 			public interface ICardUpgrade : IRoute<CardUpgrade>
 			{
 				/// <summary>
 				/// Whether the upgrade should be done in place, without moving the card.
 				/// </summary>
 				bool IsInPlace { get; set; }
+				
+				/// <summary>
+				/// A strategy that will be used to apply an upgrade to a card. Defaults to <c>null</c>, which will copy the value of the <see cref="Card.upgrade"/> field.
+				/// </summary>
+				IInPlaceCardUpgradeStrategy? InPlaceCardUpgradeStrategy { get; set; }
 
 				/// <summary>
 				/// Sets <see cref="IsInPlace"/>.
@@ -32,6 +38,41 @@ public partial interface IKokoroApi
 				/// <param name="value">The new value.</param>
 				/// <returns>This object after the change.</returns>
 				ICardUpgrade SetIsInPlace(bool value);
+				
+				/// <summary>
+				/// Sets <see cref="InPlaceCardUpgradeStrategy"/>.
+				/// </summary>
+				/// <param name="value">The new value.</param>
+				/// <returns>This object after the change.</returns>
+				ICardUpgrade SetInPlaceCardUpgradeStrategy(IInPlaceCardUpgradeStrategy? value);
+			}
+
+			public interface IInPlaceCardUpgradeStrategy
+			{
+				void ApplyInPlaceCardUpgrade(IApplyInPlaceCardUpgradeArgs args);
+
+				public interface IApplyInPlaceCardUpgradeArgs
+				{
+					/// <summary>
+					/// The game state.
+					/// </summary>
+					State State { get; }
+					
+					/// <summary>
+					/// The route that applies the upgrade to the card.
+					/// </summary>
+					CardUpgrade Route { get; }
+					
+					/// <summary>
+					/// The card to apply the upgrade to.
+					/// </summary>
+					Card TargetCard { get; }
+					
+					/// <summary>
+					/// The card chosen by the player as a template for the upgrade.
+					/// </summary>
+					Card TemplateCard { get; }
+				}
 			}
 		}
 	}