+++ Kokoro/Features/InPlaceCardUpgrade.cs
@@ -15,6 +15,14 @@ partial class ApiImplementation
 		{
 			public IKokoroApi.IV2.IInPlaceCardUpgradeApi.ICardUpgrade ModifyCardUpgrade(CardUpgrade route)
 				=> new InPlaceCardUpgradeManager.CardUpgradeWrapper(Mutil.DeepCopy(route));
+			
+			internal sealed class InPlaceCardUpgradeStrategyApplyInPlaceCardUpgradeArgs : IKokoroApi.IV2.IInPlaceCardUpgradeApi.IInPlaceCardUpgradeStrategy.IApplyInPlaceCardUpgradeArgs
+			{
+				public State State { get; internal set; } = null!;
+				public CardUpgrade Route { get; internal set; } = null!;
+				public Card TargetCard { get; internal set; } = null!;
+				public Card TemplateCard { get; internal set; } = null!;
+			}
 		}
 	}
 }
@@ -38,7 +46,19 @@ internal sealed class InPlaceCardUpgradeManager
 		if (card is null)
 			return true;
 
-		card.upgrade = newCard.upgrade;
+		if (ModEntry.Instance.Helper.ModData.GetOptionalModData<IKokoroApi.IV2.IInPlaceCardUpgradeApi.IInPlaceCardUpgradeStrategy>(__instance, "InPlaceCardUpgradeStrategy") is { } strategy)
+			ModEntry.Instance.ArgsPool.Do<ApiImplementation.V2Api.InPlaceCardUpgradeApi.InPlaceCardUpgradeStrategyApplyInPlaceCardUpgradeArgs>(args =>
+			{
+				args.State = g.state;
+				args.Route = __instance;
+				args.TargetCard = card;
+				args.TemplateCard = newCard;
+
+				strategy.ApplyInPlaceCardUpgrade(args);
+			});
+		else
+			card.upgrade = newCard.upgrade;
+
 		return false;
 	}
 	
@@ -53,10 +73,22 @@ internal sealed class InPlaceCardUpgradeManager
 			get => ModEntry.Instance.Helper.ModData.GetModDataOrDefault<bool>(route, "IsInPlace");
 			set => ModEntry.Instance.Helper.ModData.SetModData(route, "IsInPlace", value);
 		}
+		
+		public IKokoroApi.IV2.IInPlaceCardUpgradeApi.IInPlaceCardUpgradeStrategy? InPlaceCardUpgradeStrategy
+		{
+			get => ModEntry.Instance.Helper.ModData.GetOptionalModData<IKokoroApi.IV2.IInPlaceCardUpgradeApi.IInPlaceCardUpgradeStrategy>(route, "InPlaceCardUpgradeStrategy");
+			set => ModEntry.Instance.Helper.ModData.SetOptionalModData(route, "InPlaceCardUpgradeStrategy", value);
+		}
 
 		public IKokoroApi.IV2.IInPlaceCardUpgradeApi.ICardUpgrade SetIsInPlace(bool value)
 		{
 			this.IsInPlace = value;
+			return this;
+		}
+
+		public IKokoroApi.IV2.IInPlaceCardUpgradeApi.ICardUpgrade SetInPlaceCardUpgradeStrategy(IKokoroApi.IV2.IInPlaceCardUpgradeApi.IInPlaceCardUpgradeStrategy? value)
+		{
+			this.InPlaceCardUpgradeStrategy = value;
 			return this;
 		}
 	}